* Included changes from *old* b1_4/b_cray
    *  OBD_FL_DEBUG_CHECK for consistent obdecho end-to-end checking

    *  struct echo_obd cleanup

    *  obdecho support for "difficult" reply testing (test_setattr)

    *  replaced cowardly eap_from_cookie() usage with EAP_FROM_COOKIE()
       which asserts it is being called correctly

    *  bracketed echo_client_iocontrol() guts with unlock/lock_kernel()
       for increased concurrency during echo ioctl execution

    *  better thread control and stats in obd.c

    *  n-threads-over-m-objects test_brw support (required for
       obdfilter-survey)

* openibnal changes

    *  re-enabled openib on 2.5/2.6 builds
       configure --with-openib=no can be used to disable openib explicitly if
       required.

    *  fixed some 64 bit casts

    *  just use kernel's page_to_phys()

* fixed CERROR format problem in libcfs/module.c

+++ lnet/klnds/openiblnd/openiblnd.c
@@ -1289,7 +1289,8 @@ kibnal_api_startup (nal_t *nal, ptl_pid_t requested_pid,
         /*****************************************************/
 
         for (i = 0; i < IBNAL_N_SCHED; i++) {
-                rc = kibnal_thread_start (kibnal_scheduler, (void *)i);
+                rc = kibnal_thread_start (kibnal_scheduler,
+                                          (void *)((unsigned long)i));
                 if (rc != 0) {
                         CERROR("Can't spawn openibnal scheduler[%d]: %d\n",
                                i, rc);
